panfrost: Add support for 64 bit gpu_id
Update the gallium driver's use of gpu_id to support 64 bit gpu_ids.

Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/40921>

+++ src/gallium/drivers/panfrost/pan_blend_cso.h
@@ -41,7 +41,7 @@ struct panfrost_blend_state {
 };
 
 struct pan_blend_shader_cache {
-   unsigned gpu_id;
+   uint64_t gpu_id;
    uint32_t gpu_variant;
    struct pan_pool *bin_pool;
    struct hash_table *shaders;
@@ -58,7 +58,7 @@ struct pan_blend_shader {
 uint64_t panfrost_get_blend(struct panfrost_batch *batch, unsigned rt);
 
 void pan_blend_shader_cache_init(struct pan_blend_shader_cache *cache,
-                                 unsigned gpu_id, uint32_t gpu_variant,
+                                 uint64_t gpu_id, uint32_t gpu_variant,
                                  struct pan_pool *bin_pool);
 
 void pan_blend_shader_cache_cleanup(struct pan_blend_shader_cache *cache);
